コントローラー層テストの責務統一とファイル分割
